Move some lower-level helpers out of libs/backend and into libs/utils
