+++ libs/backend/src/cast_vote_records/build_cast_vote_record.ts
@@ -27,9 +27,8 @@ import {
   YesNoContest,
   YesNoVote,
 } from '@votingworks/types';
-import { getMarkStatus } from '@votingworks/utils';
+import { getContestsForBallotPage, getMarkStatus } from '@votingworks/utils';
 
-import { getContestsForBallotPage } from './page_layouts';
 import {
   ContestOptionPositionMap,
   ElectionOptionPositionMap,
