+++ libs/backend/src/cast_vote_records/index.ts
@@ -5,7 +5,5 @@ export {
 } from './build_cast_vote_record';
 export { buildCastVoteRecordReportMetadata } from './build_report_metadata';
 export * from './export';
-export * from './helpers';
 export * from './legacy_export';
 export * from './legacy_import';
-export { getContestsForBallotPage } from './page_layouts';
